Closes #64, implementation with extra prob. check
Implemented probability_accept for Serbian and English. For Serbian it rejects dem. and aug. if rand value is under certain treshold. This implementation is different from "Suggested implementation", and better as it gives better control.
Fixed adding attributes to derived words as the attributes were only copied from original word.